Agent_Based_Samples/document_generator/src/solution_accelerators/document_generator/evaluation/requirements.txt
azure-identity==1.20.0
azure-keyvault-secrets==4.7.0
PyYAML==6.0.1
semantic-kernel[azure]==1.32.2
azure-ai-projects==1.0.0b11
azure-ai-evaluation==1.8.0
azure-ai-ml==1.25.0
azureml-core==1.59.0.post1
azureml-fsspec==1.3.1
opencensus-ext-azure==1.1.14
opencensus==0.11.4
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
azure-monitor-opentelemetry==1.6.4
pandas==2.2.3